backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
requests
slowapi
firebase-admin
python-jose